add clip/coca readme
