+++ paddlemix/examples/clip/README.md
@@ -1,0 +1,203 @@
+# CLIP
+
+## 1. 模型简介
+
+[CLIP](https://arxiv.org/abs/2103.00020): Contrastive Language-Image Pre-Training, Paddle实现版本.
+
+语言-图像预训练对比学习，在很多语言、图像应用场景表现出性能优势，有着广泛的应用。
+
+<p align="center">
+  <img src="https://github.com/openai/CLIP/blob/main/CLIP.png?raw=true" align="middle" width = "800" />
+</p>
+
+注：图片引用自[openai/CLIP](https://github.com/openai/CLIP).
+
+
+### CLIP Model zoo
+
+(待补充)
+
+## 2. 环境准备
+
+1） 安装PaddleNLP develop版本
+
+```
+pip install --pre --upgrade paddlenlp -f https://www.paddlepaddle.org.cn/whl/paddlenlp.html
+```
+
+2）安装环境依赖包
+
+```
+pip install -r requirements.txt
+```
+
+3）安装`FusedLayerNorm`，在`paddlemix/external_ops/`目录下，安装fusedln包。(CLIP模型可选)
+
+```
+# 安装fusedln到python环境
+python setup.py install --prefix=$INSTALL_DIR
+# 添加安装路径到系统环境路径中
+export $PATH=$PATH:$INSTALL_DIR
+```
+
+4）安装paddlemix
+
+```
+git clone git@github.com:PaddlePaddle/PaddleMIX.git
+cd PaddleMix
+python setup.py install
+```
+
+## 3. 数据准备
+
+1) coco数据
+
+数据部分，默认使用`coco_karpathy`数据，使用该数据不需另外配置，会自动下载。解析部分参考`coco_clip.py`文件。
+
+如果想手动下载，请点击[DownLoadCoCo](https://bj.bcebos.com/v1/paddlenlp/datasets/paddlemix/coco.tar)
+
+2) 自定义数据
+
+如果需要自定义数据，推荐沿用`coco_karpathy`数据格式处理自己的数据。其中每条数据标注格式示例为:
+```
+{'caption': 'A woman wearing a net on her head cutting a cake. ', 'image': 'val2014/COCO_val2014_000000522418.jpg', 'image_id': 'coco_522418'}
+```
+更多可参考数据集中的`annotations/coco_karpathy_train.json`文件。
+
+## 4. 使用说明
+
+我们在Paddle中实现了`CLIP`的`ViT`系列模型，包括`Vit-L-14`.
+
+### 4.1 训练
+
+训练时使用`paddlemix/examples/clip/run_pretrain_dist.py`程序进行训练。
+
+训练命令及参数配置示例：
+
+这里示例采用单机8卡程序，sharding_degree=8.
+
+注意如果采用分布式策略，分布式并行关系有：`nnodes * nproc_per_node == tensor_parallel_degree * sharding_parallel_degree * dp_parallel_degree`，其中`dp_parallel_degree`参数根据其他几个值计算出来，因此需要保证`nnodes * nproc_per_node >= tensor_parallel_degree * sharding_parallel_degree`.
+
+```
+MODEL_NAME="paddlemix/CLIP/Vit-L-14/"
+IN_1K_DIR=[YOUR ImageNet1K val data path]
+
+python -m paddle.distributed.launch --nproc_per_node 8 run_pretrain_dist.py \
+    --dataloader_num_workers=2 \
+    --per_device_train_batch_size 16 \
+    --per_device_eval_batch_size 8 \
+    --model ${MODEL_NAME}  \
+    --warmup_steps 2000 \
+    --learning_rate 5e-4 \
+    --weight_decay 0.05 \
+    --adam_beta1 0.9  \
+    --adam_beta2 0.999  \
+    --adam_epsilon 1e-8  \
+    --max_grad_norm 5.0 \
+    --num_train_epochs 200 \
+    --tensor_parallel_degree 1 \
+    --sharding_parallel_degree 8 \
+    --sharding "stage2" \
+    --bf16 False \
+    --output_dir "./output" \
+    --logging_steps 1 \
+    --do_train \
+    --disable_tqdm True \
+    --save_steps 50000 \
+    --local_loss true \
+    --gather_with_grad true \
+    --pretrained_text_model ${MODEL_NAME} \
+    --classification_eval ${IN_1K_DIR} \
+
+```
+
+
+```
+# 参数说明
+
+--model #设置实际使用的模型,示例'paddlemix/CLIP/Vit-L-14/'
+
+--dataloader_num_workers #数据加载线程数量
+
+--per_device_train_batch_size #训练时单卡batch_size
+
+--learning_rate 5e-4 #global默认学习率，优先级低于visual_lr、text_lr
+
+--weight_decay 0.05 #global默认weight decay, 优先级低于visual_wd、text_wd
+
+--adam_beta1 0.9  #optimizer中beta1参数
+
+--adam_beta2 0.999  #optimizer中beta2参数
+
+--adam_epsilon 1e-8  #optimizer中epsilon参数
+
+--max_grad_norm 5.0 #最大梯度裁剪，将norm大于该值的grad裁剪到该值
+
+--num_train_epochs 200 #整体训练epoch次数
+
+--tensor_parallel_degree 1 #模型并行系数，设置为N则进行N卡间模型并行
+
+--sharding_parallel_degree 8 #显存优化策略，默认stage1，详情参考 [《ZeRO: Memory Optimizations Toward Training Trillion Parameter Models》]（https://arxiv.org/abs/1910.02054）
+
+--sharding "stage1" #显存优化策略stage选择，目前支持stage1、stage2
+
+--fp16 False #是否开启float16训练
+
+--output_dir "./output" #模型存储路径
+
+--logging_steps 1 #logging显示间隔steps
+
+--do_train #执行训练
+
+--save_steps 50000 #每多少个steps保存一次模型
+
+--local_loss true #loss中是否开启local loss
+
+--gather_with_grad true #loss中是否打开gather_with_grad
+
+--pretrained_text_model Vit-L-14 #预提取text features的模型
+
+--classification_eval ${IN_1K_DIR} #IN_1K测试数据路径
+```
+
+### 4.2 评估
+
+评估时使用`paddlemix/examples/evaclip/run_zero_shot_eval.py`程序进行评估。
+
+评估命令及参数配置示例：
+
+```
+MODEL_NAME="paddlemix/CLIP/Vit-L-14"
+
+IN_1K_DIR=[YOUR ImageNet1K val data path]
+
+python paddlemix/examples/clip/run_zero_shot_eval.py \
+    --per_device_eval_batch_size 32 \
+    --dataloader_num_workers=2 \
+    --model ${MODEL_NAME}  \
+    --fp16 False \
+    --pretrained_text_model Vit-L-14 \
+    --classification_eval ${IN_1K_DIR} \
+    --output_dir "output" \
+    --disable_tqdm True \
+```
+
+```
+# 参数说明
+
+--model #设置实际使用的模型,示例'paddlemix/CLIP/Vit-L-14'
+
+--dataloader_num_workers #数据加载线程数量
+
+--per_device_eval_batch_size #评估时单卡batch_size
+
+--fp16 False #是否开启fp16推理
+
+--pretrained_text_model Vit-L-14 #预提取text features的模型
+
+--classification_eval ${IN_1K_DIR} #IN_1K测试数据路径
+
+--output_dir "output" #模型输出文件路径
+
+--disable_tqdm True #是否关闭tqdm进度条
+```
